feat(update) Parse command line
Run into a problem. A command can be specified that has spaces by
quoting. The command will appear as a single option to the top level $@
variable. However when $@ is passed to lower level functions, the shell
appears to be separating the arguements by white space. I don't believe
that this should be happening. My solution is write this code in ruby
instead.